+++ src/Playground/Benchmark/ZoneTree2.cs
@@ -16,7 +16,7 @@ public class ZoneTree2
         Console.WriteLine("WriteAheadLogMode: = " + mode);
 
         var dataPath = "../../data/TestStringTree" + mode + count;
-        if (Directory.Exists(dataPath))
+        if (TestConfig.RecreateDatabases && Directory.Exists(dataPath))
             Directory.Delete(dataPath, true);
 
         var stopWatch = new Stopwatch();
