+++ src/ZoneTree/WAL/LazyFileSystemWriteAheadLog.cs
@@ -28,10 +28,16 @@ public sealed class LazyFileSystemWriteAheadLog<TKey, TValue> : IWriteAheadLog<T
         ISerializer<TKey> keySerializer,
         ISerializer<TValue> valueSerializer,
         string filePath,
-        int compressionBlockSize)
+        int compressionBlockSize,
+        bool enableTailWriterJob,
+        int tailWriterJobInterval)
     {
         FilePath = filePath;
-        FileStream = new CompressedFileStream(filePath, compressionBlockSize);
+        FileStream = new CompressedFileStream(
+            filePath,
+            compressionBlockSize,
+            enableTailWriterJob,
+            tailWriterJobInterval);
         FileStream.Seek(0, SeekOrigin.End);
         KeySerializer = keySerializer;
         ValueSerializer = valueSerializer;
@@ -92,7 +98,11 @@ public sealed class LazyFileSystemWriteAheadLog<TKey, TValue> : IWriteAheadLog<T
         Queue.Clear();
         StopWriter(false);
         FileStream.Dispose();
-        File.Delete(FilePath);
+        if (File.Exists(FilePath))
+            File.Delete(FilePath);
+        var tailPath = FilePath + ".tail";
+        if (File.Exists(tailPath))
+            File.Delete(tailPath);
     }
 
     struct LogEntry
